Move code from horace_nxs_inst.py to eniius/horace.py
Fix bugs in to_icp()
Readded mcstas_nxs_inst.py to generate nxs from instr
Add .gitignore
