+++ eniius/mcstas.py
@@ -477,48 +477,3 @@ class NXMcStas():
         return nxinst
 
 
-def conv_types(obj):
-    typ = type(obj)
-    dtyp = np.dtype(typ)
-    if dtyp.name == 'object':
-        if isinstance(obj, np.ndarray):
-            vl = obj.tolist()
-            el = vl[0]
-            for ii in range(1, len(obj.shape)):
-                el = el[0]
-            tp = np.dtype(type(el)).name
-        elif obj is None:
-            (tp, vl) = ('string', 'None')
-        elif isinstance(obj, nexus.NXattr):
-            val = obj.nxdata
-            if isinstance(val, np.ndarray):
-                val = val.tolist()
-            if obj.dtype == 'object':
-                (tp, vl) = (np.dtype(type(obj.nxdata)).name, val)
-            else:
-                (tp, vl) = (obj.dtype, val)
-        else:
-            raise RuntimeError(f'unrecognised type {typ} / {dtyp}')
-    else:
-        (tp, vl) = (dtyp.name, obj)
-    if tp == 'str':
-        tp = 'string'
-    elif tp == 'float64':
-        tp = 'float'
-    elif tp == 'object':
-        raise RuntimeError('Internal logical error')
-    return tp, vl
-
-
-def mcstas2nxs(instrfile):
-    instobj = get_instr(instrfile)
-    nxinst = NXMcStas(instobj.component_list).NXinstrument()
-    nxinst['name'] = NXfield(value=instobj.name)
-    from .writer import Writer
-    writer = Writer(nxinst)
-    writer.to_json(f'mcstas_{instobj.name}.json')
-    writer.to_icp(f'mcstas_{instobj.name}.nxs')
-
-
-if __name__ == '__main__':
-    mcstas2nxs(sys.argv[1])
