+++ eniius/writer.py
@@ -1,6 +1,7 @@
 import numpy as np
 import nexusformat.nexus
 import warnings
+import json
 
 VERSION = '0.1'
 
@@ -16,10 +17,9 @@ def new_wg(self, group):
             self[path].attrs['NX_class'] = np.array(group.nxclass, dtype='S')
     return links
 def new_wd(self, data):
-    nxpath = f'{self.nxpath}/{data.nxname}'
     if data._target is None and data._uncopied_data is None and data._memfile is None \
             and (not data.nxfile or data.nxfile.filename == self.filename) \
-            and data.is_string() and isinstance(data._value, str):
+            and data.dtype is not None and data.is_string():
         data._shape = (1,)
         data._dtype = f'S{len(data._value)}'
         data._value = np.array(data._value, dtype='S')
@@ -27,8 +27,41 @@ def new_wd(self, data):
 nexusformat.nexus.tree.NXFile._writegroup = new_wg
 nexusformat.nexus.tree.NXFile._writedata = new_wd
 
-
 from nexusformat.nexus import *
+
+
+def conv_types(obj):
+    typ = type(obj)
+    dtyp = np.dtype(typ)
+    if dtyp.name == 'object':
+        if isinstance(obj, np.ndarray):
+            vl = obj.tolist()
+            el = vl[0]
+            for ii in range(1, len(obj.shape)):
+                el = el[0]
+            tp = np.dtype(type(el)).name
+        elif obj is None:
+            (tp, vl) = ('string', 'None')
+        elif isinstance(obj, NXattr):
+            val = obj.nxdata
+            if isinstance(val, np.ndarray):
+                val = val.tolist()
+            if obj.dtype == 'object':
+                (tp, vl) = (np.dtype(type(obj.nxdata)).name, val)
+            else:
+                (tp, vl) = (obj.dtype, val)
+        else:
+            raise RuntimeError(f'unrecognised type {typ} / {dtyp}')
+    else:
+        (tp, vl) = (dtyp.name, obj)
+    if tp == 'str':
+        tp = 'string'
+    elif tp == 'float64':
+        tp = 'float'
+    elif tp == 'object':
+        raise RuntimeError('Internal logical error')
+    return tp, vl
+
 
 class Writer:
     """
@@ -66,10 +99,12 @@ class Writer:
             self.nxobj[self.rootname] = nxentry
 
 
-    def to_json(self, outfile):
+    def to_json(self, outfile, nxobj=None):
         # Converts a NeXus object to a JSON-compatible dictionary
         children = []
-        for k, obj in self.nxobj.items():
+        if nxobj is None:
+            nxobj = self.nxobj
+        for k, obj in nxobj.items():
             if hasattr(obj, 'nxclass'):
                 if obj.nxclass == 'NXfield':
                     typ, val = conv_types(obj.nxdata)
@@ -80,7 +115,7 @@ class Writer:
                     entry = {'name':k, 'type':'group'}
                     attrs = [{'name':'NX_class', 'dtype':'string', 'values':obj.nxclass}]
                     if len(obj._entries) > 0:
-                        entry['children'] = nexus2jsondict(obj)
+                        entry['children'] = self.to_json(outfile, obj)
             else:
                 raise RuntimeError(f'unrecognised object key {k}')
             for n, v in obj.attrs.items():
@@ -95,7 +130,7 @@ class Writer:
             f.write(json.dumps({'children':children}, indent=4))
 
 
-    def to_nxspe(self, outfile, ei=25):
+    def to_nxspe(self, outfile, ei=25, det_file=None):
         if not outfile.endswith('.nxspe'):
             outfile += '.nxspe'
         with nxopen(outfile, 'w') as root:
@@ -104,10 +139,10 @@ class Writer:
             root['w1/NXSPE_info'] = NXcollection(fixed_energy=NXfield(ei, units='meV'),
                                                            ki_over_kf_scaling=True,
                                                            psi=NXfield(np.nan, units='degrees'))
-            root['w1/instrument'] = self.instr
+            root['w1/instrument'] = self.inst
             if self.sample is not None:
                 root['w1/sample'] = self.sample
-            if self.sample is None:
+            if self.data is None:
                 n = 5
                 dmat = np.random.rand(n, n)
                 emat = np.random.rand(n, n) / 10.
@@ -122,6 +157,9 @@ class Writer:
                 root['w1/data'] = dat
             else:
                 root['w1/data'] = self.data
+            if det_file:
+                for ky, val in self._parse_det(det_file).items():
+                    root[f'w1/{ky}'] = val
 
 
     def to_icp(self, outfile, det_file=None):
@@ -131,27 +169,31 @@ class Writer:
             root['mantid_workspace_1'] = NXentry()
             root['mantid_workspace_1/program_name'] = NXfield('eniius', version=VERSION)
             root['mantid_workspace_1/instrument'] = self.inst
-            inst_name = root['mantid_workspace_1/instrument/name'].nxvalue + ' '
-            root['mantid_workspace_1/instrument/name'].replace(
-                    NXfield(np.array(inst_name, dtype='S'), shape=(1,)))
             if det_file:
-                # Assumes ISIS format; cols=(det#, delta, L2, code, theta, phi, W_xyz, a_xyz, det_123)
-                with open(det_file, 'r') as f:
-                    titles = [next(f) for x in range(3)][2].split()
-                    if titles[0] == 'det' and titles[1] == 'no':
-                        titles = titles[1:]
-                    titles = ','.join(titles[6:])
-                detdat = np.loadtxt(det_file, skiprows=3)
-                for fnm, idn in zip(['detectors', 'monitors'], [2, 1]):
-                    idx = np.where(detdat[:,3] == idn)[0]
-                    fd = {f'number_of_{fnm}': NXfield(np.array(len(idx), dtype='uint64'))}
-                    fd['detector_number'] = NXfield(detdat[idx,0].astype(np.int32))
-                    fd['detector_offset'] = NXfield(detdat[idx,1])
-                    fd['distance'] = NXfield(detdat[idx,2], units='metre')
-                    fd['polar_angle'] = NXfield(detdat[idx,4], units='degree')
-                    fd['azimuthal_angle'] = NXfield(detdat[idx,5], units='degree')
-                    fd['user_table_titles'] = NXfield(np.array(titles, dtype='S'), shape=(1,))
-                    for j in range(6, detdat.shape[1]):
-                        fd[f'user_table_{j-5}'] = NXfield(detdat[idx,j])
-                    assert len(titles.split(',')) == detdat.shape[1]-6, "Number of titles not commensurate with table width"
-                    root['mantid_workspace_1/instrument/physical_' + fnm] = NXdetector(**fd)
+                for ky, val in self._parse_det(det_file).items():
+                    root[f'mantid_workspace_1/{ky}'] = val
+
+
+    def _parse_det(self, det_file):
+        # Assumes ISIS format; cols=(det#, delta, L2, code, theta, phi, W_xyz, a_xyz, det_123)
+        with open(det_file, 'r') as f:
+            titles = [next(f) for x in range(3)][2].split()
+            if titles[0] == 'det' and titles[1] == 'no':
+                titles = titles[1:]
+            titles = ','.join(titles[6:])
+        detdat = np.loadtxt(det_file, skiprows=3)
+        rv = {}
+        for fnm, idn in zip(['detectors', 'monitors'], [2, 1]):
+            idx = np.where(detdat[:,3] == idn)[0]
+            fd = {f'number_of_{fnm}': NXfield(np.array(len(idx), dtype='uint64'))}
+            fd['detector_number'] = NXfield(detdat[idx,0].astype(np.int32))
+            fd['detector_offset'] = NXfield(detdat[idx,1])
+            fd['distance'] = NXfield(detdat[idx,2], units='metre')
+            fd['polar_angle'] = NXfield(detdat[idx,4], units='degree')
+            fd['azimuthal_angle'] = NXfield(detdat[idx,5], units='degree')
+            fd['user_table_titles'] = NXfield(titles)
+            for j in range(6, detdat.shape[1]):
+                fd[f'user_table_{j-5}'] = NXfield(detdat[idx,j])
+            assert len(titles.split(',')) == detdat.shape[1]-6, "Number of titles not commensurate with table width"
+            rv['instrument/physical_' + fnm] = NXdetector(**fd)
+        return rv
